refactor: move Buff into GameLogic
