+++ server/src/GameLogic/Buff/OffensiveBuff.cs
@@ -1,7 +1,4 @@
-using Serilog.Debugging;
-using Thuai.Server.GameLogic;
-
-namespace Thuai.Server.Buff;
+namespace Thuai.Server.GameLogic.Buff;
 
 public static class O_Buff
 {
